% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcall_barplot.R
\name{process_vcalls}
\alias{process_vcalls}
\title{process_vcalls}
\usage{
process_vcalls(vcall_tbl, vcall_col = "V_CALL")
}
\arguments{
\item{vcall_tbl}{tibble of data}

\item{vcall_col}{name of the column that contains the v calls [default: "V_CALL"]}
}
\value{
tibble of raw counts and percentages
}
\description{
Counts occurrences of a V_CALLs in a tibble. Returns a tibble of raw counts and percentages
}
\examples{
processed_vcalls <- process_vcalls(vcall_eg)

}
