% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcall_barplot.R
\name{vcall_barplot}
\alias{vcall_barplot}
\title{vcall_barplot}
\usage{
vcall_barplot(vcall_data, x_val = "VCALL", y_val = "percentage")
}
\arguments{
\item{vcall_data}{output from process_vcalls}

\item{x_val}{what to plot on x axis, "VCALL" [default]}

\item{y_val}{what to plot on y axis, "percentage" [default] or "raw_count"}
}
\value{
ggplot object
}
\description{
vcall_barplot
}
\examples{
processed_vcalls <- process_vcalls(vcall_eg)
vcall_barplot(processed_vcalls)
}
