% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcall_eg.R
\docType{data}
\name{vcall_eg}
\alias{vcall_eg}
\title{Output of pipeline}
\format{
A tibble with 2000 rows and x columns:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\usage{
vcall_eg
}
\description{
A truncated example dataset containing V_CALLs and other output from pipeline.
}
\keyword{datasets}
